+++ src/main/java/org/example/spring_meuble/dao/FurnitureRepository.java
@@ -3,5 +3,5 @@ package org.example.spring_meuble.dao;
 import org.example.spring_meuble.entity.Furniture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FurnitureRepository extends JpaRepository<Furniture, Long> {
+public interface FurnitureRepository extends JpaRepository<Furniture, Integer> {
 }
